working multithreading TODO synchronization

+++ src/block/BlockFactory.java
@@ -2,7 +2,7 @@ package block;
 
 public class BlockFactory {
 
-    public static Block createBlock(String prevHash, int requiredNumberOfZeros, int minerID) {
+    public static Block createBlock(String prevHash, int requiredNumberOfZeros, long minerID) {
         HashInfo hashInfo = HashFactory.generateHash(prevHash, requiredNumberOfZeros, minerID);
         return new Block(hashInfo);
     }
